JS-5542: space settings wip
