+++ src/json/text.json
@@ -524,7 +524,7 @@
 	"pageSettingsColorModeButtonDark": "Dark Mode",
 	"pageSettingsColorModeButtonSystem": "Auto Mode",
 
-	"pageSettingsSpaceOverview": "Overview",
+	"pageSettingsSpaceGeneral": "General",
 	"pageSettingsSpaceRemoteStorage": "Remote Storage",
 	"pageSettingsSpaceIntegrations": "Integrations",
 	"pageSettingsSpaceManageContent": "Manage Content",
